Refactory add new domain layer
