+++ VacationRental.Api/Controllers/CalendarController.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using Microsoft.AspNetCore.Mvc;
+using VacationRental.Api.Domain;
 using VacationRental.Api.Models;
 
 namespace VacationRental.Api.Controllers
@@ -11,6 +12,7 @@ namespace VacationRental.Api.Controllers
     {
         private readonly IDictionary<int, RentalViewModel> _rentals;
         private readonly IDictionary<int, BookingViewModel> _bookings;
+        private Calendar _calendarDomain;
 
         public CalendarController(
             IDictionary<int, RentalViewModel> rentals,
@@ -18,6 +20,7 @@ namespace VacationRental.Api.Controllers
         {
             _rentals = rentals;
             _bookings = bookings;
+            _calendarDomain = new Calendar();
         }
 
         [HttpGet]
@@ -28,37 +31,7 @@ namespace VacationRental.Api.Controllers
             if (!_rentals.ContainsKey(rentalId))
                 throw new ApplicationException("Rental not found");
 
-            return CreateCalendarResponse(rentalId, start, nights);
-        }
-
-        private CalendarViewModel CreateCalendarResponse(int rentalId, DateTime start, int nights)
-        {
-            var resultCalendar = new CalendarViewModel
-            {
-                RentalId = rentalId,
-                Dates = new List<CalendarDateViewModel>()
-            };
-            for (var i = 0; i < nights; i++)
-            {
-                var date = new CalendarDateViewModel
-                {
-                    Date = start.Date.AddDays(i),
-                    Bookings = new List<CalendarBookingViewModel>()
-                };
-
-                foreach (var booking in _bookings.Values)
-                {
-                    if (booking.RentalId == rentalId
-                        && booking.Start <= date.Date && booking.Start.AddDays(booking.Nights) > date.Date)
-                    {
-                        date.Bookings.Add(new CalendarBookingViewModel { Id = booking.Id });
-                    }
-                }
-
-                resultCalendar.Dates.Add(date);
-            }
-
-            return resultCalendar;
+            return _calendarDomain.CreateCalendarResponse(rentalId, start, nights, _bookings);
         }
     }
 }
