+++ VacationRental.Api/VacationRental.Api.csproj
@@ -13,4 +13,8 @@
     <PackageReference Include="Swashbuckle.AspNetCore.SwaggerUI" Version="4.0.1" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Domain\Rental\" />
+  </ItemGroup>
+
 </Project>
